requirements.txt
llama-index==0.10.18
llama-index-core==0.10.21.post1
llama-index-readers-file==0.1.9
llama-index-embeddings-huggingface==0.1.4
python-dotenv==1.0.1
tiktoken==0.6.0
weaviate-client<4.0.0
llama-index-vector-stores-weaviate==0.1.4
llama-index-embeddings-openai==0.1.6
llama-index-llms-openai==0.1.7
openai
sentence-transformers
streamlit==1.34.0
fpdf
